requirements.txt
Django==3.2.17
django-dsfr==0.13.1
python-dotenv==0.21.0
pre-commit==2.20.0
flake8==5.0.4
black==22.8.0
isort==5.10.1
psycopg2-binary==2.9.5
mozilla-django-oidc==3.0.0
django-referrer-policy==1.0
django-csp==3.7
whitenoise==6.2.0
dj-database-url==1.2.0
gunicorn==20.1.0
requests==2.28.2
sib-api-v3-sdk==7.5.0
factory-boy==3.2.1
